requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.2
faiss-cpu==1.13.2
groq==0.4.1
httpx<0.27
pydantic==2.5.0
pydantic-settings==2.1.0
